Update checkboxes ramps
