+++ src/components/checkbox-group/checkbox-group-component.jsx
@@ -2,7 +2,6 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import Checkbox from './checkbox';
-import styles from './checkbox-group-styles.module.scss';
 
 const CheckboxGroup = ({ options, theme, checkedOptions, handleClick }) => {
   const handleChange = (e, option) => {
@@ -13,10 +12,10 @@ const CheckboxGroup = ({ options, theme, checkedOptions, handleClick }) => {
   }
 
   return (
-    <div className={styles.container}>
+    <div>
       {options.map(option => {
         return (
-          <Checkbox key={option.name} option={option} theme={theme[option.theme]} checked={checkedOptions[option.value]} onChange={handleChange} />
+          <Checkbox key={option.name} option={option} theme={theme} checked={checkedOptions[option.value]} onChange={handleChange} />
         )
       })}
     </div>
