--- src/components/checkbox-group/checkbox-group-styles.module.scss
@@ -1,1 +1,0 @@
-@import 'styles/settings';
