--- src/components/protected-areas-layers/protected-areas-layers-styles.module.scss
@@ -1,15 +1,0 @@
-@import 'styles/settings';
-
-.overrideCheckbox-protected-areas {
-  input[type="checkbox"]:checked + label::before {
-    background: $persimmon;
-    border: 2px solid $white; 
-  }
-}
-
-.overrideCheckbox-community-areas {
-  input[type="checkbox"]:checked + label::before {
-    background: $saffron-mango;
-    border: 2px solid $white; 
-  }
-}
